requirements.txt
asgiref
Django
sqlparse
gunicorn
django-heroku
dj-database-url
django-allauth
requests
PyJWT
cryptography
whitenoise
django-environ
django-storages
django-debug-toolbar
boto3